Split ItemMegaMenu into container & presentational - #4639
