+++ ui/app/src/components/ItemMegaMenu/ItemMegaMenuUI.tsx
@@ -1,0 +1,310 @@
+/*
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Menu from '@material-ui/core/Menu';
+import MenuItem from '@material-ui/core/MenuItem';
+import MenuList from '@material-ui/core/MenuList';
+import EmptyState from '../SystemStatus/EmptyState';
+import { FormattedMessage } from 'react-intl';
+import clsx from 'clsx';
+import Typography from '@material-ui/core/Typography';
+import ItemDisplay from '../ItemDisplay';
+import ItemPublishingTargetIcon from '../ItemPublishingTargetIcon';
+import ItemStateIcon from '../ItemStateIcon';
+import { getItemPublishingTargetText, getItemStateText } from '../ItemDisplay/utils';
+import React, { ReactNode } from 'react';
+import { CSSProperties } from '@material-ui/styles';
+import {
+  createStyles,
+  makeStyles,
+  PopoverOrigin,
+  PopoverPosition,
+  PopoverProps,
+  PopoverReference
+} from '@material-ui/core';
+import palette from '../../styles/palette';
+import { SystemIconDescriptor } from '../SystemIcon';
+import { DetailedItem } from '../../models/Item';
+import { ContextMenuOption } from '../ContextMenu';
+import GlobalState from '../../models/GlobalState';
+
+export type ItemMegaMenuUIClassKey =
+  | 'root'
+  | 'menuMainList'
+  | 'mainItem'
+  | 'actionsContainer'
+  | 'actionsColumn'
+  | 'emptyRoot'
+  | 'itemsList'
+  | 'itemInfo'
+  | 'itemInfoContentType'
+  | 'itemEdited'
+  | 'itemEditedText'
+  | 'itemState'
+  | 'infoItem'
+  | 'menuItem'
+  | 'itemTypeIcon'
+  | 'itemTypography'
+  | 'icon';
+
+export type ItemMegaMenuUIStyles = Partial<Record<ItemMegaMenuUIClassKey, CSSProperties>>;
+
+export interface MenuOption {
+  id: string;
+  icon?: SystemIconDescriptor;
+  label: ReactNode;
+}
+
+export interface ItemMegaMenuUIProps {
+  open: boolean;
+  styles?: ItemMegaMenuUIStyles;
+  classes?: Partial<Record<ItemMegaMenuUIClassKey, string>>;
+  item: DetailedItem;
+  options: ContextMenuOption[][];
+  editorialOptions: ContextMenuOption[];
+  nonEditorialOptions: ContextMenuOption[][];
+  anchorEl?: PopoverProps['anchorEl'];
+  anchorOrigin?: PopoverOrigin;
+  anchorReference?: PopoverReference;
+  anchorPosition?: PopoverPosition;
+  contentType: string;
+  locale: GlobalState['uiConfig']['locale'];
+  onClose?(): void;
+  onMenuItemClicked(option: string, event: React.MouseEvent<Element, MouseEvent>): void;
+}
+
+export const useStyles = makeStyles((theme) =>
+  createStyles<ItemMegaMenuUIClassKey, ItemMegaMenuUIStyles>({
+    root: (styles) => ({
+      borderRadius: '12px',
+      ...styles.root
+    }),
+    menuMainList: (styles) => ({
+      padding: 0,
+      ...styles.menuMainList
+    }),
+    mainItem: (styles) => ({
+      padding: '10px 20px',
+      ...styles.mainItem
+    }),
+    actionsContainer: (styles) => ({
+      display: 'flex',
+      flexDirection: 'row',
+      width: '100%',
+      ...styles.actionsContainer
+    }),
+    actionsColumn: (styles) => ({
+      display: 'flex',
+      flexDirection: 'column',
+      flexBasis: '100%',
+      flex: '1',
+      '&:first-child': {
+        marginRight: '60px'
+      },
+      ...styles.actionsColumn
+    }),
+    emptyRoot: (styles) => ({
+      display: 'block',
+      padding: '10px',
+      textAlign: 'center',
+      ...styles.emptyRoot
+    }),
+    itemsList: (styles) => ({
+      padding: 0,
+      ...styles.itemsList
+    }),
+    itemInfo: (styles) => ({
+      display: 'block',
+      borderBottom: `1px solid ${palette.gray.light4}`,
+      ...styles.itemInfo
+    }),
+    itemInfoContentType: (styles) => ({
+      color: theme.palette.text.secondary,
+      marginBottom: '4px',
+      ...styles.itemInfoContentType
+    }),
+    itemEdited: (styles) => ({
+      paddingTop: '12px',
+      borderTop: `1px solid ${palette.gray.light4}`,
+      ...styles.itemEdited
+    }),
+    itemEditedText: (styles) => ({
+      color: theme.palette.text.secondary,
+      fontWeight: 600,
+      ...styles.itemEditedText
+    }),
+    itemState: (styles) => ({
+      '&> *': {
+        marginRight: '5px'
+      },
+      ...styles.itemState
+    }),
+    infoItem: (styles) => ({
+      cursor: 'default',
+      '&:hover': {
+        backgroundColor: 'inherit'
+      },
+      ...styles.infoItem
+    }),
+    menuItem: (styles) => ({
+      paddingLeft: 0,
+      paddingRight: 0,
+      minWidth: '100px',
+      ...styles.menuItem
+    }),
+    itemTypeIcon: (styles) => ({
+      color: palette.teal.main,
+      marginRight: '2px',
+      ...styles.itemTypeIcon
+    }),
+    itemTypography: (styles) => ({
+      color: theme.palette.text.primary,
+      ...styles.itemTypography
+    }),
+    icon: (styles) => ({
+      fontSize: '0.8rem',
+      verticalAlign: 'middle',
+      ...styles.icon
+    })
+  })
+);
+
+export default function ItemMegaMenuUI(props: ItemMegaMenuUIProps) {
+  const {
+    open,
+    styles,
+    item,
+    options,
+    editorialOptions,
+    nonEditorialOptions,
+    anchorEl,
+    anchorOrigin,
+    anchorReference,
+    anchorPosition,
+    contentType,
+    locale,
+    onClose,
+    onMenuItemClicked,
+    classes: propClasses
+  } = props;
+  const classes = useStyles(styles);
+
+  return (
+    <Menu
+      open={open}
+      onClose={onClose}
+      anchorEl={anchorEl}
+      anchorOrigin={anchorOrigin}
+      anchorReference={anchorReference}
+      anchorPosition={anchorPosition}
+      classes={{
+        paper: classes.root,
+        list: classes.menuMainList,
+        ...propClasses
+      }}
+    >
+      {options.flatMap((i) => i).length === 0 ? (
+        <EmptyState
+          title={
+            <FormattedMessage id="contextMenu.emptyOptionsMessage" defaultMessage="No options available to display." />
+          }
+        />
+      ) : (
+        [
+          <MenuItem key={0} className={clsx(classes.itemInfo, classes.infoItem, classes.mainItem)}>
+            <Typography variant="body2" component="h2" className={classes.itemInfoContentType}>
+              {contentType}
+            </Typography>
+            <ItemDisplay
+              item={item}
+              showPublishingTarget={false}
+              showWorkflowState={false}
+              classes={{ icon: classes.itemTypeIcon }}
+              labelTypographyProps={{
+                className: classes.itemTypography
+              }}
+            />
+            <div className={classes.itemState}>
+              <ItemPublishingTargetIcon item={item} className={classes.icon} />
+              <Typography variant="body2" component="span">
+                {getItemPublishingTargetText(item?.stateMap)}
+              </Typography>
+              <ItemStateIcon item={item} className={classes.icon} />
+              <Typography variant="body2" component="span">
+                {getItemStateText(item?.stateMap)}
+              </Typography>
+            </div>
+          </MenuItem>,
+          <div key={1} className={clsx(classes.mainItem, classes.actionsContainer)}>
+            <MenuList className={clsx(classes.actionsColumn, classes.itemsList)}>
+              {editorialOptions.map((option: MenuOption, y: number) => (
+                <MenuItem
+                  dense
+                  key={option.id}
+                  onClick={(e) => onMenuItemClicked(option.id, e)}
+                  className={clsx(classes.menuItem, propClasses?.menuItem)}
+                  children={option.label}
+                />
+              ))}
+            </MenuList>
+            <div className={classes.actionsColumn}>
+              {nonEditorialOptions.map((section: any, i: number) => (
+                <MenuList key={i} className={classes.itemsList}>
+                  {section.map((option: MenuOption, y: number) => (
+                    <MenuItem
+                      dense
+                      key={option.id}
+                      divider={i !== nonEditorialOptions.length - 1 && y === section.length - 1}
+                      onClick={(e) => onMenuItemClicked(option.id, e)}
+                      className={clsx(classes.menuItem, propClasses?.menuItem)}
+                      children={option.label}
+                    />
+                  ))}
+                </MenuList>
+              ))}
+            </div>
+          </div>,
+          <MenuItem key={2} className={clsx(classes.itemEdited, classes.infoItem, classes.mainItem)}>
+            <Typography variant="body2">
+              <FormattedMessage
+                id="itemMegaMenu.editedBy"
+                defaultMessage="{edited} {date} {byLabel} {by}"
+                values={{
+                  date: new Intl.DateTimeFormat(locale.localeCode, locale.dateTimeFormatOptions).format(
+                    new Date(item?.sandbox.dateModified)
+                  ),
+                  by: item?.sandbox.modifier,
+                  edited: (
+                    <span className={classes.itemEditedText}>
+                      <FormattedMessage id="words.edited" defaultMessage="Edited" />
+                    </span>
+                  ),
+                  byLabel: item?.sandbox.modifier ? (
+                    <span className={classes.itemEditedText}>
+                      <FormattedMessage id="words.by" defaultMessage={'by'} />
+                    </span>
+                  ) : (
+                    ''
+                  )
+                }}
+              />
+            </Typography>
+          </MenuItem>
+        ]
+      )}
+    </Menu>
+  );
+}
